+++ classes/class.ilHelpMePlugin.php
@@ -4,10 +4,11 @@ require_once __DIR__ . "/../vendor/autoload.php";
 if (file_exists(__DIR__ . "/../../../../Cron/CronHook/HelpMeCron/vendor/autoload.php")) {
 	require_once __DIR__ . "/../../../../Cron/CronHook/HelpMeCron/vendor/autoload.php";
 }
-require_once __DIR__ . "/../../Notifications4Plugins/vendor/autoload.php";
 
 use srag\DIC\HelpMe\Util\LibraryLanguageInstaller;
 use srag\Plugins\HelpMe\Config\Config;
+use srag\Plugins\HelpMe\Notification\Notification\Language\NotificationLanguage;
+use srag\Plugins\HelpMe\Notification\Notification\Notification;
 use srag\Plugins\HelpMe\Project\Project;
 use srag\Plugins\HelpMe\Ticket\Ticket;
 use srag\Plugins\HelpMe\Utils\HelpMeTrait;
@@ -71,6 +72,9 @@ class ilHelpMePlugin extends ilUserInterfaceHookPlugin {
 
 		LibraryLanguageInstaller::getInstance()->withPlugin(self::plugin())->withLibraryLanguageDirectory(__DIR__
 			. "/../vendor/srag/custominputguis/src/ScreenshotsInputGUI/lang")->updateLanguages();
+
+		LibraryLanguageInstaller::getInstance()->withPlugin(self::plugin())->withLibraryLanguageDirectory(__DIR__
+			. "/../vendor/srag/notifications4plugin/lang")->updateLanguages();
 	}
 
 
@@ -79,6 +83,8 @@ class ilHelpMePlugin extends ilUserInterfaceHookPlugin {
 	 */
 	protected function deleteData()/*: void*/ {
 		self::dic()->database()->dropTable(Config::TABLE_NAME, false);
+		self::dic()->database()->dropTable(Notification::TABLE_NAME, false);
+		self::dic()->database()->dropTable(NotificationLanguage::TABLE_NAME, false);
 		self::dic()->database()->dropTable(Project::TABLE_NAME, false);
 		self::dic()->database()->dropTable(Ticket::TABLE_NAME, false);
 	}