+++ sql/dbupdate.php
@@ -99,3 +99,88 @@ if (\srag\DIC\HelpMe\DICStatic::dic()->database()->tableColumnExists(\srag\Plugi
 	\srag\DIC\HelpMe\DICStatic::dic()->database()->dropTableColumn(\srag\Plugins\HelpMe\Project\Project::TABLE_NAME, "project_fix_version");
 }
 ?>
+<#14>
+<?php
+\srag\Plugins\HelpMe\Notification\Notification\Notification::updateDB();
+\srag\Plugins\HelpMe\Notification\Notification\Language\NotificationLanguage::updateDB();
+
+$templates = \srag\Plugins\HelpMe\Config\Config::getField(\srag\Plugins\HelpMe\Config\Config::KEY_RECIPIENT_TEMPLATES);
+
+if (\srag\Notifications4Plugin\HelpMe\Notification\Repository::getInstance(\srag\Plugins\HelpMe\Notification\Notification\Notification::class, \srag\Plugins\HelpMe\Notification\Notification\Language\NotificationLanguage::class)
+		->migrateFromOldGlobalPlugin($templates[\srag\Plugins\HelpMe\Recipient\RecipientCreateJiraTicket::SEND_EMAIL]) === null) {
+
+	$notification = \srag\Notifications4Plugin\HelpMe\Notification\Repository::getInstance(\srag\Plugins\HelpMe\Notification\Notification\Notification::class, \srag\Plugins\HelpMe\Notification\Notification\Language\NotificationLanguage::class)
+		->factory()->newInstance();
+
+	$notification->setName($templates[\srag\Plugins\HelpMe\Recipient\RecipientCreateJiraTicket::SEND_EMAIL] = \srag\Plugins\HelpMe\Recipient\RecipientCreateJiraTicket::SEND_EMAIL);
+	$notification->setTitle("Mail");
+	$notification->setDefaultLanguage(\srag\DIC\HelpMe\DICStatic::dic()->language()->getDefaultLanguage() === "de" ? "de" : "en");
+
+	foreach ([ "de", "en" ] as $lang) {
+		$notification->setSubject("{{ support.getTitle }}", $lang);
+		$notification->setText("{% for key,value in fields %}
+<p>
+	<h2>{{ key }}</h2>
+	{{ value }}
+</p>
+<br>
+{% endfor %}", $lang);
+	}
+
+	\srag\Notifications4Plugin\HelpMe\Notification\Repository::getInstance(\srag\Plugins\HelpMe\Notification\Notification\Notification::class, \srag\Plugins\HelpMe\Notification\Notification\Language\NotificationLanguage::class)
+		->storeInstance($notification);
+}
+
+if (\srag\Notifications4Plugin\HelpMe\Notification\Repository::getInstance(\srag\Plugins\HelpMe\Notification\Notification\Notification::class, \srag\Plugins\HelpMe\Notification\Notification\Language\NotificationLanguage::class)
+		->migrateFromOldGlobalPlugin($templates[\srag\Plugins\HelpMe\Recipient\RecipientCreateJiraTicket::CREATE_JIRA_TICKET]) === null) {
+
+	$notification = \srag\Notifications4Plugin\HelpMe\Notification\Repository::getInstance(\srag\Plugins\HelpMe\Notification\Notification\Notification::class, \srag\Plugins\HelpMe\Notification\Notification\Language\NotificationLanguage::class)
+		->factory()->newInstance();
+
+	$notification->setName($templates[\srag\Plugins\HelpMe\Recipient\RecipientCreateJiraTicket::CREATE_JIRA_TICKET] = \srag\Plugins\HelpMe\Recipient\RecipientCreateJiraTicket::CREATE_JIRA_TICKET);
+	$notification->setTitle("Jira");
+	$notification->setDefaultLanguage(\srag\DIC\HelpMe\DICStatic::dic()->language()->getDefaultLanguage() === "de" ? "de" : "en");
+
+	foreach ([ "de", "en" ] as $lang) {
+		$notification->setSubject("{{ support.getTitle }}", $lang);
+		$notification->setText("{% for key,value in fields %}
+{{ key }}:
+{{ value }}
+
+
+{% endfor %}", $lang);
+	}
+
+	\srag\Notifications4Plugin\HelpMe\Notification\Repository::getInstance(\srag\Plugins\HelpMe\Notification\Notification\Notification::class, \srag\Plugins\HelpMe\Notification\Notification\Language\NotificationLanguage::class)
+		->storeInstance($notification);
+}
+
+if (\srag\Notifications4Plugin\HelpMe\Notification\Repository::getInstance(\srag\Plugins\HelpMe\Notification\Notification\Notification::class, \srag\Plugins\HelpMe\Notification\Notification\Language\NotificationLanguage::class)
+		->migrateFromOldGlobalPlugin($templates[\srag\Plugins\HelpMe\Config\Config::KEY_SEND_CONFIRMATION_EMAIL]) === null) {
+
+	$notification = \srag\Notifications4Plugin\HelpMe\Notification\Repository::getInstance(\srag\Plugins\HelpMe\Notification\Notification\Notification::class, \srag\Plugins\HelpMe\Notification\Notification\Language\NotificationLanguage::class)
+		->factory()->newInstance();
+
+	$notification->setName($templates[\srag\Plugins\HelpMe\Config\Config::KEY_SEND_CONFIRMATION_EMAIL] = \srag\Plugins\HelpMe\Config\Config::KEY_SEND_CONFIRMATION_EMAIL);
+	$notification->setTitle("Confirm Mail");
+	$notification->setDefaultLanguage(\srag\DIC\HelpMe\DICStatic::dic()->language()->getDefaultLanguage() === "de" ? "de" : "en");
+
+	foreach ([ "de", "en" ] as $lang) {
+		$notification->setSubject(\srag\DIC\HelpMe\DICStatic::plugin(\ilHelpMePlugin::class)
+				->translate("confirmation", \srag\Plugins\HelpMe\Support\SupportGUI::LANG_MODULE_SUPPORT, [], true, $lang)
+			. ": {{ support.getTitle }}", $lang);
+		$notification->setText("{% for key,value in fields %}
+<p>
+	<h2>{{ key }}</h2>
+	{{ value }}
+</p>
+<br>
+{% endfor %}", $lang);
+	}
+
+	\srag\Notifications4Plugin\HelpMe\Notification\Repository::getInstance(\srag\Plugins\HelpMe\Notification\Notification\Notification::class, \srag\Plugins\HelpMe\Notification\Notification\Language\NotificationLanguage::class)
+		->storeInstance($notification);
+}
+
+\srag\Plugins\HelpMe\Config\Config::setField(\srag\Plugins\HelpMe\Config\Config::KEY_RECIPIENT_TEMPLATES, $templates);
+?>