+++ src/Config/ConfigFormGUI.php
@@ -6,21 +6,19 @@ use ilCheckboxInputGUI;
 use ilEMailInputGUI;
 use ilHelpMePlugin;
 use ilMultiSelectInputGUI;
-use ilNotifications4PluginsPlugin;
 use ilPasswordInputGUI;
 use ilRadioGroupInputGUI;
 use ilRadioOption;
 use ilTextAreaInputGUI;
 use ilTextInputGUI;
 use srag\ActiveRecordConfig\HelpMe\ActiveRecordConfigFormGUI;
-use srag\DIC\Notifications4Plugins\DICStatic as Notifications4PluginsDICStatic;
 use srag\JiraCurl\HelpMe\JiraCurl;
-use srag\Notifications4Plugin\Notifications4Plugins\Utils\Notifications4PluginTrait;
+use srag\Notifications4Plugin\HelpMe\Utils\Notifications4PluginTrait;
+use srag\Plugins\HelpMe\Notification\Notification\Language\NotificationLanguage;
+use srag\Plugins\HelpMe\Notification\Notification\Notification;
 use srag\Plugins\HelpMe\Recipient\Recipient;
 use srag\Plugins\HelpMe\Support\Support;
 use srag\Plugins\HelpMe\Utils\HelpMeTrait;
-use srag\Plugins\Notifications4Plugins\Notification\Language\NotificationLanguage;
-use srag\Plugins\Notifications4Plugins\Notification\Notification;
 
 /**
  * Class ConfigFormGUI
@@ -186,7 +184,7 @@ class ConfigFormGUI extends ActiveRecordConfigFormGUI {
 	 * @return array
 	 */
 	protected function getTemplateSelection(string $template_name): array {
-		return self::notificationUI()->withPlugin(Notifications4PluginsDICStatic::plugin(ilNotifications4PluginsPlugin::class))
+		return self::notificationUI()->withPlugin(self::plugin())
 			->templateSelection(self::notification(Notification::class, NotificationLanguage::class)
 				->getArrayForSelection(self::notification(Notification::class, NotificationLanguage::class)
 					->getNotifications()), Config::KEY_RECIPIENT_TEMPLATES . "_" . $template_name, [