+++ src/Recipient/RecipientSendMail.php
@@ -6,7 +6,7 @@ use ilMimeMail;
 use PHPMailer\PHPMailer\Exception as phpmailerException;
 use srag\ActiveRecordConfig\HelpMe\Exception\ActiveRecordConfigException;
 use srag\DIC\HelpMe\Exception\DICException;
-use srag\Notifications4Plugin\Notifications4Plugins\Exception\Notifications4PluginException;
+use srag\Notifications4Plugin\HelpMe\Exception\Notifications4PluginException;
 use srag\Plugins\HelpMe\Config\Config;
 use srag\Plugins\HelpMe\Exception\HelpMeException;
 use srag\Plugins\HelpMe\Support\Support;