+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## [x]
+- Notifications4Plugin library
+
 ## [4.0.2]
 - Fix goto show tickets not work on ILIAS 5.3